base: fix word-size-sensitive helper signatures from the @inbounds removal
The 32-bit testers (i686-linux-gnu, i686-w64-mingw32) on buildkite
julia-pr build 242 failed on helpers introduced by the @inbounds-removal
outlining whose scalar parameters were typed `::Int` but receive values
that are `Int64` independent of the platform word size:

- `_wexp_tail` in ryu/exp.jl takes the decimal exponent, which the Ryu
  algorithm computes as `Int64`; broke every `writeexp` call on 32-bit
  (ryu and Printf test suites).
- `_diffelt` in multidimensional.jl is indexed with the indices of the
  underlying range; a `StepRange{DateTime,Day}` uses `Int64` indices.

Both are widened to `::Integer`. Also make the effect-split tests use
word-size-neutral test data, and run the generated no-throw codegen
assertions only on 64-bit platforms: their entries were generated and
verified against x86_64 codegen, whose Int-typed method signatures and
LLVM bounds-check-elision behavior 32-bit targets do not share.

This commit was written with the assistance of generative AI.

+++ Compiler/test/effect_split.jl
@@ -79,7 +79,7 @@ esplit_sum_plain(A::Vector{Int}, upto::Int) = esplit_sum_partial(A, upto)
 esplit_sum_split(A::Vector{Int}, upto::Int) = Core.invoke_split_effects(:nothrow, esplit_sum_partial, A, upto)
 
 @testset "synthesized precondition: runtime semantics" begin
-    A = collect(Int64(1):Int64(100))
+    A = collect(1:100)
     @test esplit_sum_split(A, 100) === esplit_sum_plain(A, 100)
     @test esplit_sum_split(A, 17) === esplit_sum_plain(A, 17)
     @test esplit_sum_split(A, 0) === 0
